+++ src/builtin/mod.rs
@@ -1,0 +1,5 @@
+mod functions;
+mod string_builtins;
+
+pub use functions::builtins;
+pub use string_builtins::{dot_str_builtins, DotBuiltinKind};
